+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.1.0:
 
 * fix bug caused by the removal of `mysql_database` LWRP from the mysql cookbook starting with version 1.2.0
+* fix most foodcritic warnings
 
 ## v1.0.0: